Software/requirements.txt
bottle==0.12.8
cffi==1.14.5
click==7.1.2
cycler==0.10.0
decorator==4.4.2
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.2
playsound==1.2.2
pycparser==2.20
pylab-sdk==1.3.2
pyparsing==2.4.7
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
pyrtlsdr==0.2.92
python-box==5.3.0
python-dateutil==2.8.1
python-dotenv==0.15.0
PyYAML==5.4.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
samplerate==0.1.0
scipy==1.6.1
self==2020.12.3
simpleaudio==1.0.4
six==1.15.0
